💚 CHORE: import 경로 수정, Entity 위치 변경
