+++ src/chat/chat.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { ChatService } from './chat.service';
 import { ChatController } from './chat.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ChatRoom } from '../database/entities/chatroom.entity';
-import { User } from '../database/entities/user.entity';
-import { ChatUser } from '../database/entities/chatuser.entity';
+import { ChatRoom } from '../common/entities/chatroom.entity';
+import { User } from '../common/entities/user.entity';
+import { ChatUser } from '../common/entities/chatuser.entity';
 import { ChatGateway } from './chat.gateway';
 import { RepositoryModule } from '../repository/repository.module';
