+++ src/common/enums.ts
@@ -7,3 +7,23 @@ export enum UserStatus {
 export enum EAchievement {
   TEN_WINS = 'TenWins',
 }
+
+export enum ChatRoomMode {
+  PUBLIC = 'PUBLIC',
+  PROTECTED = 'PROTECTED',
+  PRIVATE = 'PRIVATE',
+  DIRECT = 'DIRECT',
+}
+
+export enum ChatUserStatus {
+  NONE = 'NONE',
+  KICK = 'KICK',
+  BAN = 'BAN',
+  MUTE = 'MUTE',
+}
+
+export enum ChatRole {
+  USER = 'USER',
+  ADMIN = 'ADMIN',
+  OWNER = 'OWNER',
+}
