--- src/database/entities/enums.ts
@@ -1,25 +1,0 @@
-export enum UserStatus {
-  ONLINE = 'ONLINE',
-  OFFLINE = 'OFFLINE',
-  GAMING = 'GAMING',
-}
-
-export enum ChatRoomMode {
-  PUBLIC = 'PUBLIC',
-  PROTECTED = 'PROTECTED',
-  PRIVATE = 'PRIVATE',
-  DIRECT = 'DIRECT',
-}
-
-export enum ChatUserStatus {
-  NONE = 'NONE',
-  KICK = 'KICK',
-  BAN = 'BAN',
-  MUTE = 'MUTE',
-}
-
-export enum ChatRole {
-  USER = 'USER',
-  ADMIN = 'ADMIN',
-  OWNER = 'OWNER',
-}
